requirements.txt
# Core dependencies
lxml>=4.9.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.66.0

# Async
aiohttp>=3.9.0
aiodns>=3.0.0

# Graph database
neo4j>=5.14.0

# Citation qualification (Sprint 4)
rank-bm25>=0.2.2
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# LLM Classification (Sprint 6)
openai>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
